requirements.txt
scikit-learn==1.3.0
numpy==1.26.4
matplotlib==3.7.0
##torch==2.0.1
reformer-pytorch==1.4.4
mamba-ssm==1.2.0
pandas==1.4.0